api: refactor stream utilities #171
